[feat] font 수정 완료
