+++ app/src/main/java/com/omoolen/omooroid/home/fragments/one/event/DeadlineEvent.kt
@@ -1,6 +1,6 @@
-package com.omoolen.omooroid.home.fragments.one.networkApi
+package com.omoolen.omooroid.home.fragments.one.event
 
-data class LastestEvent(
+data class DeadlineEvent(
     val content: String,
     val id: Int,
     val image: String,
