+++ app/src/main/java/com/omoolen/omooroid/home/fragments/one/networkApi/ResponseOneList.kt
@@ -1,16 +1,22 @@
 package com.omoolen.omooroid.home.fragments.one.networkApi
 
 import com.google.gson.annotations.SerializedName
+import com.omoolen.omooroid.home.fragments.one.curating.FindRecomendationByUser
+import com.omoolen.omooroid.home.fragments.one.event.DeadlineEvent
+import com.omoolen.omooroid.home.fragments.one.event.LastestEvent
+import com.omoolen.omooroid.home.fragments.one.recommend.RecommendationBySeason
+import com.omoolen.omooroid.home.fragments.one.recommend.RecommendationBySituation
+import com.omoolen.omooroid.home.fragments.one.tip.Guide
 
 data class ResponseOneList(
 
     @SerializedName("data")
-    val `data`: ResponesOneData,
+    val `data`: ResponseOneData,
     val message: String,
     val status: Int,
     val succses: Boolean
 )
-data class ResponesOneData(
+data class ResponseOneData(
 
     val deadlineEvent: List<DeadlineEvent>,
     val findRecomendationByUser: List<FindRecomendationByUser>,
