+++ app/src/main/java/com/omoolen/omooroid/home/fragments/one/recommend/RecommendationBySituation.kt
@@ -1,6 +1,6 @@
-package com.omoolen.omooroid.home.fragments.one.networkApi
+package com.omoolen.omooroid.home.fragments.one.recommend
 
-data class FindRecomendationByUser(
+data class RecommendationBySituation(
     val brand: String,
     val category: String,
     val changeCycle: Int,
