+++ app/src/main/res/drawable/home_indicator_unselected.xml
@@ -1,15 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item>
-
-        <shape android:shape="oval">
-            <solid android:color="@color/om_fourth_gray" />
-            <stroke
-                android:width="4dp"
-                android:color="@android:color/transparent" />
-            <size
-                android:width="8dp"
-                android:height="8dp" />
-        </shape>
-    </item>
-</layer-list>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+    android:shape="oval">
+    <stroke android:color="@android:color/transparent" android:width="13dp"/>
+    <solid
+        android:color="@color/om_fourth_gray"/>
+    <size
+        android:width="17dp"
+        android:height="17dp"/>
+</shape>
