+++ app/src/main/res/layout/item_one_curating.xml
@@ -38,6 +38,7 @@
             android:layout_height="wrap_content"
             android:layout_marginTop="14dp"
             tools:text="오렌즈"
+            android:includeFontPadding="false"
             android:text="@{curatingInfo.brand}"
             android:theme="@style/TextStyle15"
             app:layout_constraintStart_toStartOf="parent"
@@ -48,6 +49,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="4dp"
+            android:includeFontPadding="false"
             android:text="@{curatingInfo.name}"
             android:theme="@style/TextStyle16"
             app:layout_constraintStart_toStartOf="parent"
@@ -59,6 +61,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:layout_marginTop="4dp"
+            android:includeFontPadding="false"
             android:text="@{curatingInfo.info}"
             android:theme="@style/TextStyle17"
             app:layout_constraintStart_toStartOf="parent"
@@ -72,6 +75,7 @@
             android:layout_height="wrap_content"
             android:layout_marginTop="8dp"
             android:theme="@style/TextStyle18"
+            android:includeFontPadding="false"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@+id/tv_one_recommend_info"
             tools:text="18,000원" />
