+++ app/src/main/res/layout/item_one_new.xml
@@ -66,6 +66,7 @@
                 android:layout_marginStart="13dp"
                 android:layout_marginBottom="2dp"
                 android:text="@{newInfo.brand}"
+                android:includeFontPadding="false"
                 android:theme="@style/TextStyle22"
                 app:layout_constraintBottom_toTopOf="@+id/tv_one_new_name"
                 app:layout_constraintStart_toEndOf="@+id/iv_one_new_circle"
@@ -79,6 +80,7 @@
                 android:layout_marginBottom="20dp"
                 android:text="@{newInfo.name1}"
                 android:theme="@style/TextStyle23"
+                android:includeFontPadding="false"
                 app:layout_constraintBottom_toBottomOf="@+id/iv_one_new"
                 app:layout_constraintStart_toEndOf="@+id/iv_one_new_circle"
                 tools:text="제품 상세 정보 이름" />
@@ -119,6 +121,7 @@
                         android:layout_marginTop="5dp"
                         android:text="@{newInfo.name1}"
                         android:theme="@style/TextStyle24"
+                        android:includeFontPadding="false"
                         app:layout_constraintStart_toEndOf="@+id/iv_one_new_len1"
                         app:layout_constraintTop_toBottomOf="@+id/tv_one_new_brand_name1"
                         tools:text="브라운 컬러 익스 렌즈" />
@@ -131,6 +134,7 @@
                         android:layout_marginTop="12dp"
                         android:text="@{newInfo.brand}"
                         android:theme="@style/TextStyle21"
+                        android:includeFontPadding="false"
                         app:layout_constraintStart_toEndOf="@+id/iv_one_new_len1"
                         app:layout_constraintTop_toTopOf="parent"
                         tools:text="오렌즈" />
@@ -144,6 +148,7 @@
                         android:layout_marginTop="6dp"
                         android:textColor="@color/om_main_black"
                         android:theme="@style/TextStyle25"
+                        android:includeFontPadding="false"
                         app:layout_constraintStart_toEndOf="@+id/iv_one_new_len1"
                         app:layout_constraintTop_toBottomOf="@+id/tv_one_new_name1"
                         tools:text="18,000원" />
@@ -189,6 +194,7 @@
                         android:layout_marginTop="5dp"
                         android:text="@{newInfo.name2}"
                         android:theme="@style/TextStyle24"
+                        android:includeFontPadding="false"
                         app:layout_constraintStart_toEndOf="@+id/iv_one_new_len2"
                         app:layout_constraintTop_toBottomOf="@+id/tv_one_new_brand_name2"
                         tools:text="브라운 컬러 익스 렌즈" />
@@ -202,6 +208,7 @@
                         android:text="@{newInfo.brand}"
                         android:textColor="@color/om_main_black"
                         android:theme="@style/TextStyle21"
+                        android:includeFontPadding="false"
                         app:layout_constraintStart_toEndOf="@+id/iv_one_new_len2"
                         app:layout_constraintTop_toTopOf="parent"
                         tools:text="오렌즈" />
@@ -215,6 +222,7 @@
                         android:layout_marginTop="6dp"
                         android:textColor="@color/om_main_black"
                         android:theme="@style/TextStyle25"
+                        android:includeFontPadding="false"
                         app:layout_constraintStart_toEndOf="@+id/iv_one_new_len2"
                         app:layout_constraintTop_toBottomOf="@+id/tv_one_new_name2"
                         tools:text="18,000원" />
@@ -259,6 +267,7 @@
                         android:layout_marginTop="5dp"
                         android:text="@{newInfo.name3}"
                         android:theme="@style/TextStyle24"
+                        android:includeFontPadding="false"
                         app:layout_constraintStart_toEndOf="@+id/iv_one_new_len3"
                         app:layout_constraintTop_toBottomOf="@+id/tv_one_new_brand_name3"
                         tools:text="브라운 컬러 익스 렌즈" />
@@ -272,6 +281,7 @@
                         android:text="@{newInfo.brand}"
                         android:textColor="@color/om_main_black"
                         android:theme="@style/TextStyle21"
+                        android:includeFontPadding="false"
                         app:layout_constraintStart_toEndOf="@+id/iv_one_new_len3"
                         app:layout_constraintTop_toTopOf="parent"
                         tools:text="오렌즈" />
@@ -283,6 +293,7 @@
                         android:layout_height="wrap_content"
                         android:layout_marginStart="16dp"
                         android:layout_marginTop="6dp"
+                        android:includeFontPadding="false"
                         android:textColor="@color/om_main_black"
                         android:theme="@style/TextStyle25"
                         app:layout_constraintStart_toEndOf="@+id/iv_one_new_len3"
