+++ app/src/main/res/layout/item_one_tip.xml
@@ -43,6 +43,7 @@
         android:layout_marginTop="26dp"
         android:text="@string/home_tip_sub"
         android:theme="@style/TextStyle53"
+        android:includeFontPadding="false"
         app:layout_constraintStart_toStartOf="@+id/iv_one_tip_header"
         app:layout_constraintTop_toTopOf="@+id/iv_one_tip_header" />
 
@@ -52,11 +53,9 @@
         android:layout_height="wrap_content"
         android:layout_marginStart="16dp"
         android:layout_marginTop="2dp"
-        android:letterSpacing="-0.9"
         android:text="이런이런 정보가 들어가요!"
-        android:textColor="@color/om_white"
-        android:textSize="18sp"
-        android:textStyle="bold"
+        android:includeFontPadding="false"
+        android:theme="@style/TextStyle73"
         app:layout_constraintStart_toStartOf="@+id/iv_one_tip_header"
         app:layout_constraintTop_toBottomOf="@+id/tv_one_tip_sub" />
 
@@ -70,6 +69,7 @@
         android:text="@{tipInfo.title1}"
         tools:text="랜즈 눈치 안보고 패션을 꾸밀 수 있다?"
         android:theme="@style/TextStyle19"
+        android:includeFontPadding="false"
         app:layout_constraintStart_toStartOf="@+id/iv_one_tip_bg"
         app:layout_constraintTop_toBottomOf="@+id/iv_one_tip_header" />
 
@@ -82,6 +82,7 @@
         android:text="@{tipInfo.sub1}"
         tools:text="1분 렌즈 상식에 대한 간단 설명"
         android:theme="@style/TextStyle20"
+        android:includeFontPadding="false"
         app:layout_constraintStart_toStartOf="@+id/iv_one_tip_bg"
         app:layout_constraintTop_toBottomOf="@+id/tv_one_tip_title1" />
 
@@ -104,6 +105,7 @@
         android:text="@{tipInfo.title2}"
         tools:text="랜즈 눈치 안보고 패션을 꾸밀 수 있다?"
         android:theme="@style/TextStyle19"
+        android:includeFontPadding="false"
         app:layout_constraintStart_toStartOf="@+id/iv_one_tip_bg"
         app:layout_constraintTop_toBottomOf="@+id/iv_line1" />
 
@@ -116,6 +118,7 @@
         android:text="@{tipInfo.sub2}"
         tools:text="1분 렌즈 상식에 대한 간단 설명"
         android:theme="@style/TextStyle20"
+        android:includeFontPadding="false"
         app:layout_constraintStart_toStartOf="@+id/iv_one_tip_bg"
         app:layout_constraintTop_toBottomOf="@+id/tv_one_tip_title2" />
 
@@ -138,6 +141,7 @@
         android:text="@{tipInfo.title3}"
         tools:text="랜즈 눈치 안보고 패션을 꾸밀 수 있다?"
         android:theme="@style/TextStyle19"
+        android:includeFontPadding="false"
         app:layout_constraintStart_toStartOf="@+id/iv_one_tip_bg"
         app:layout_constraintTop_toBottomOf="@+id/iv_line2" />
 
@@ -149,6 +153,7 @@
         android:layout_marginTop="4dp"
         android:text="@{tipInfo.sub3}"
         android:theme="@style/TextStyle20"
+        android:includeFontPadding="false"
         tools:text="1분 렌즈 상식에 대한 간단 설명"
         app:layout_constraintStart_toStartOf="@+id/iv_one_tip_bg"
         app:layout_constraintTop_toBottomOf="@+id/tv_one_tip_title3" />
